+++ src/shared-components/index.js
@@ -3,5 +3,6 @@ export { default as Alert } from './alert';
 export { default as Button, RoundButton } from './button';
 export { default as Container } from './container';
 export { default as RadioButton } from './radioButton';
+export { default as Checkbox } from './checkbox';
 export { default as Chip } from './chip';
 export { default as Typography, style as TYPOGRAPHY_STYLE } from './typography';